refactor: atoms 관련 파일 수정
